requirements.txt
<<<<<<< HEAD
pandas
PyYAML
scikit-learn
python-dotenv
loguru
pyarrow
fastAPI
uvicorn
=======
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
cramjam==2.9.1
    # via fastparquet
fastparquet==2024.11.0
    # via mise-en-prod-data-science (pyproject.toml)
fsspec==2025.2.0
    # via fastparquet
joblib==1.4.2
    # via scikit-learn
loguru==0.7.3
    # via mise-en-prod-data-science (pyproject.toml)
numpy==2.2.3
    # via
    #   fastparquet
    #   pandas
    #   scikit-learn
    #   scipy
packaging==24.2
    # via fastparquet
pandas==2.2.3
    # via
    #   mise-en-prod-data-science (pyproject.toml)
    #   fastparquet
pyarrow==19.0.1
    # via mise-en-prod-data-science (pyproject.toml)
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via mise-en-prod-data-science (pyproject.toml)
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via mise-en-prod-data-science (pyproject.toml)
scikit-learn==1.6.1
    # via mise-en-prod-data-science (pyproject.toml)
scipy==1.15.2
    # via scikit-learn
six==1.17.0
    # via python-dateutil
threadpoolctl==3.5.0
    # via scikit-learn
tzdata==2025.1
    # via pandas
>>>>>>> 2010a02e160a40b150f10aac3ab597014e4064ee
